Added api to fetch all posts
There is now an api (api/sauron/posts) that fetches all posts and sends them as a json array. Useful.

Removed useless link from visible page in /index

Made posts more modular ! The code for a post can now be reused. It still needs to be spoon fed the data, but that's the responsibility of the page.
That means toilet reviews will be easier to implement.

Renamed post_generals.module.css to just post.module.css