Added Containers, and some navbar changes
Adds two basic containers, rowContainer and columnContainer.
This decision was made because if one were to put a navbar, for example,
in a page, you would need to know in advance what the navbar react component
ACTUALLY returns. In this case, in the current state of things, it returns
a <nav> with a <div> alongside it if it's on mobile. The unpredictability
of the returned component begs for encapsulation and isolation ; containers
provide exactly that. Right now, with this commit, to use a component, all
one needs to know is whether the component expects a row or a column type of
display (generally a column, as would be the case in standard html)

Made navbar a bit more responsive (still looks horrible/
doesn't work on desktop) by using different logic for the two. Mobile
has a popup, desktop has a menu.

Extracted some post uility function into their own file.

Added some css so that the __next div takes as much space as possible.
Not sure if this one is actually useful, but it doesn't hurt or do anything.

Changed the navtest canary page so that it could compile because of some apostrophes.